partial fix tdf100961 doc import: E_Locked field as plain text
The date/time fields should be a const time,
not the current, updateable time.

This patch throws away the field and just imports
the plain text (and paves the way for locking
the field as well).

A better fix requires a better programmer,
one who can take that plain text string and
use the formatting information to
set the DateTime field to that time.

But at least this plain-text date will be better
than a completely wrong field date.

Change-Id: I8a57ecde64aa670b2205fcf99ed9ae226579c4ed
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/113986
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Justin Luth <justin_luth@sil.org>