IsInput is equal to IsInterrupt, so drop it
Nobody is really interested in the reason for the interrupt.

And since we already checked IsInterrupt() to get in this
branch, mbFormatContentOnInterrupt is always true. Also because
we want to finish layouting the current page, especially on Idle,
according to commit aa355eeec8c5526c68b17d15192aef94062607d7
("#i42586# ... - finish format of current page on interrupt."),
so must ignore many (all?) IsInterrupt() checks.

Change-Id: Ic1d576130a605348a86115ebeb1f016d8163c6f0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114095
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>