tdf#134951 docxexport: stop duplicating stuff in postponed text
Certain attributes like comments, fields, and footnotes
should not be duplicated when a character run is split
by the bPostponeWritingText hack.

Only DOCX PostponesWritingText.

The comment-with-range is still not perfect,
but at least it is not fully duplicated.
Best would be to just get rid of the hack?

A few more items suggested themselves as belonging
to this list, but I don't really know what they are,
and no instances of them in this position were found
in the existing unit tests. So I just left them as a
commented possibility.

Change-Id: I950c4250b5f7ed62d63eeff71cabaa10b67dca5f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114148
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>