tdf#94326 doc import: allow direct numbering to cancel outlineStyle
Prior to this, if a Chapter Numbering style was assigned
as the paragraph style, any direct numbering would have
been skipped.

In this unit test, DOC defines WW8Num2 as the
numbering style for the Heading styles.
LO copies this to its special "Outline" style.
So any direct formatting with WW8Num2 must be ignored.
However, any other numbering rules should be allowed,
and that is what this patch does.

Make sure that WW8Num2 can also be applied outside of
the Heading styles. (Unfortunately, the numbering
is probably broken since these are now considered
to be two different lists - unless someone has
connected them in the background some way.)

Change-Id: If29c8bbd096acc0edd6f1fe32cedc054f16460d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114298
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>