Revert "tdf#125949 Allow the slide to continue with ..." (tdf#133447)
These animations were causing problems because of Cairo not handling
well large matrix values. With the real problem fixed this can now
be reverted. I assume the commit claiming to occur with OpenGL was
a mistake, I cannot reproduce any problem, or possibly the problem
went away when the OpenGL VCL backend was removed.

This reverts commit 8eb2d2972583b909a249f5b0f22a9b1fbf533d24.
This reverts commit 3e88fc6b0eef06e1e12fcfe765e3092c6c06ce5c.

Change-Id: I1c1fff94ff65c7937197a32176a8775ecb2a502e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114345
Tested-by: Luboš Luňák <l.lunak@collabora.com>
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>