tdf#106541 doc import: always register listLevel
LFO and listLevel are NOT a required pair.

The path on the Paragraph side is safe to re-enter.
This patch makes the style side safe to re-run
with partial information.

If we don't register immediately, then a non-existant LFO
means that we totally lose the provided listLevel.
So just RegisterNumFormat even though it
likely will be run again by Read_LFOPosition.

Change-Id: Iad10cdd1ff7e6fc0a77b01849b3b157703945556
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114400
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>