tdf#90101 add .uno:PasteAsLink
No attributes or notes are pasted, i.e. only the links.
Links to dates loose the date formatting currently, see tdf#142093.

Paste As Link is only added as command.
No entries to menus or context menus are made
as this functionality is not widely used, see tdf#102255

Icons are added with tdf#141885

Change-Id: Ifbdaa076045b5f10073a658115fea211bb30e6a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/115137
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>