tdf#141966 writerfilter CN: fix chapter number identification
This totally up-ends the current way of doing things,
but it was so totally broken, and confused us for a very long time.
Chapter Numbering has nothing to do with "Heading X" styles!!!!!!!!
Thus, we can clean up a lot of nonsense clauses.

The existing code makes it sound like the constraints of
"outline numbering" are an MS format thing, but actually it is
just an implementation detail of how LO contrains styles
and listLevels.

Yes, it is still nice to be able to round-trip the
Chapter Numbering settings, so we won't ignore it completely,
but really, all the complexity comes from trying to cram
the normal flexibility of inheritable, style-assignable
listLevels into a single special-purpose numbering rule.

One thing that has always killed us is that direct formatting
could not share in this unique style. But it needs to in DOCX.
So, the key here is to NAME-ASSOCIATE the "Outline" style,
so that ALL references to that numId end up pointing to the
same rule, even if it is the super-special oddball one.

P.S. Chapter Numbering should have NO IMPACT on the
import process. It should only affect UI interaction
when a user assigns a paragraph style: whether LO can
apply a numbering listLevel or not.

THIS IS EXTREMELY DANGEROUS TERRITORY! WAIT FOR 7.3!

existing unit tests that mapped to a numId other than 1:
ooxmlexport3: tdf95495 - numId 4
ooxmlexport4: tdf81345 - numId 3
ooxmlexport5: tdf123262 - numId 48
ooxmlexport8: fdo74745 - numId 6
ooxmlexport10: tdf89165 - numId 12
ooxmlexport14: tdf133605 - numId 2
ooxmlexport16: tdf132752 - numId 2

and one example which didn't take the first available choice:
ooxmlexport13: tdf95848 - preferred numId 3 over numId 2

Change-Id: I561cd1594f198ad402893c77c1c983f206f53c57
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/115614
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>