Distinguish between --disable-gui and --without-x
For the static WASM build, we need to build with GUI but without
X11, as this just includes a static Qt backend.

We already have USING_X11, which indicates that the platform will
use X11 as a backend, so we can match --without-x to USING_X11=no,
independent from the --disable-gui setting.

Fitting all this into vcl/Library_vcl.mk was originally getting
quite complex, but by moving the logic into configure.ac in a
previous commit, the branching becomes way easier.

Change-Id: Ia2ddba3400b4306b609a1f64823cca75061e0593
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116125
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>