gtk4: restore GtkFixed container for SalObjects
put the DrawingArea into an Overlay so the overlay will take the size of
the DrawingArea. Put a GtkFixed into that overlay which will then also
mirror the size of the DrawingArea. Keep that GtkFixed because the
SalObject stuff is set up to use a GtkFixed and its just easier to do it
this way than use the overlay directly.

Change-Id: I937b3740fd809660ee0edef56d5cf036f2503892
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116059
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>