tdf#141683 fix missing flag in multi sel pasted transposed as links
For transpose, the links for option 'as links' will be set during
TransposeClip(). Thus, formulas will be created. So, the formula flag must be
set for subsequent copying to the document. This flag must be set in the multi
range selection code path for transposed as links.

I've adapted the tests. Unfortunately, this fix cannot be tested directly with
automatic tests. However, the same behavior is applied in the unit tests.

Change-Id: Ia16a8371894a784fc7fcfa1c7b5242f552bef956
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116332
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>